refactor: migrate all usage handling to typed EndpointUsage
This commit completely replaces the untyped JSON usage handling with a strongly typed EndpointUsage system:
1. Remove old UsageAccumulator and ToOpenAIUsage traits along with related helper functions
2. Update all usage-related structs to use i64 instead of i32/u64 for token counts
3. Replace raw JSON usage fields in UnifiedResponse and UnifiedEvent with typed EndpointUsage
4. Simplify usage parsing and conversion across all provider and HTTP mapping code
5. Remove redundant usage aggregation logic in agent loop and tracing code
6. Standardize on typed usage payloads for all endpoints

+++ src/llm_provider/vertexai/types.rs
@@ -131,17 +131,17 @@ pub struct VertexFunctionResponse {
 #[serde(rename_all = "camelCase")]
 pub struct VertexUsageMetadata {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub prompt_token_count: Option<i32>,
+    pub prompt_token_count: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub candidates_token_count: Option<i32>,
+    pub candidates_token_count: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub cached_content_token_count: Option<i32>,
+    pub cached_content_token_count: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub tool_use_prompt_token_count: Option<i32>,
+    pub tool_use_prompt_token_count: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub thoughts_token_count: Option<i32>,
+    pub thoughts_token_count: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub total_token_count: Option<i32>,
+    pub total_token_count: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub cache_tokens_details: Option<Vec<VertexModalityTokenCount>>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -160,7 +160,7 @@ pub struct VertexModalityTokenCount {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub modality: Option<VertexMediaModality>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub token_count: Option<i32>,
+    pub token_count: Option<i64>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
